1.1 - Begin work on implementing Misc decks.
The JapaneseWebView allows misc type decks to view the front of the card and back notes (when answered correctly). This allows the user to add web content to the front of the card like with Anki.

Working on implementing the Misc Deck card editor, not fully implemted yet. It will allow users to store images and generate TTS audio or select audio file for use as part of the question.