1.0.4 - Disable undo button on start of review, add Jitai-like functionality
The "Cycle through fonts" option in Advanced Preferences allow users to view the world for review, lessons, and item view with a list of fonts, providing functionality like the "Jitai" extension in WaniKani.

Undo button should only enable when user gets the word wrong when Anki mode is not enabled.